Extract join notification to service
Move the logic that finds other players' FCM tokens and sends a "player joined" notification into a new notifyOtherPlayersOfJoin helper in server/services/notificationService.js. Update server/controller/match.js to call the new helper (with try/catch) instead of performing the DB query and send inline, so notification errors are logged and won't break the join response. Also import the User model in the notification service and export the new function.
